Move action_log_for_game out of nn_evaluator
